Menu management

+++ src/restaurant/menu/menu.service.ts
@@ -9,11 +9,10 @@ import { StatusMessage } from "../enums/status-message.enum";
 import { UpdateMenuCategoryDto } from "./dto/update.menu-category.dto";
 import { CreateMenuDto } from "./dto/create.menu.dto";
 import { UpdateMenuDto } from "./dto/update.menu.dto";
-import * as fs from 'fs';
 
-import * as base64js from 'base64-js';
 import { MenuManagementEvents } from "src/gateway/dto";
-import { deleteFile, fileExist } from "src/utils";
+import { deleteFile, fileExist, getRootDirectory } from "src/utils";
+import { response } from "express";
 
 
 @Injectable()
@@ -30,11 +29,9 @@ export class MenuService {
         }
       },
     });
-
     if (category) {
       throw new BadRequestException(StatusMessage.Exist);
     }
-
     //Add new category
     const status = dto.status.toString().toLowerCase() == 'true' ? true : false;
     const menuCategory = await this.prisma.menuCategory.create({
@@ -63,16 +60,7 @@ export class MenuService {
         }
       ]
     }));
-    for (let i = 0; i < categories.length; i++) {
-      try {
-        const base64Image = await this.ConvertToBase64String(categories[i].image)
-        categories[i].image = base64Image
-      } catch (error) {
-        categories[i].image = 'https://upload.wikimedia.org/wikipedia/commons/thumb/6/6d/Good_Food_Display_-_NCI_Visuals_Online.jpg/1280px-Good_Food_Display_-_NCI_Visuals_Online.jpg'
-      }
-
-    }
-    return new APIResponse(Status.Success, StatusMessage.GetSuccess, categories);
+    return categories;
   }
 
   async getRestaurantMenuCategoryById(restaurantId: string, categoryId: string) {
@@ -85,9 +73,9 @@ export class MenuService {
         }
       });
       if (!category) {
-        return new APIResponse(Status.OtherErrors, StatusMessage.NoRecord, null);
+        throw new NotFoundException(StatusMessage.NoRecord);
       }
-      return new APIResponse(Status.Success, StatusMessage.GetSuccess, category);
+      return category;
     } catch (error) {
       throw error
     }
@@ -135,7 +123,7 @@ export class MenuService {
       throw new BadRequestException(StatusMessage.Exist)
     }
 
-    if (await fileExist(category.image)) {
+    if (file && await fileExist(category.image)) {
       await deleteFile(category.image)
     }
 
@@ -148,7 +136,7 @@ export class MenuService {
       data: {
         name: dto.name,
         description: dto.description,
-        image: file.path,
+        image: file ? file.path : category.image,
         restaurantId: parseInt(restaurantId),
         deleted: false,
         status
@@ -205,7 +193,7 @@ export class MenuService {
   }
 
   async getRestaurantMenu(restaurantId: string) {
-    const menus = (await this.prisma.menu.findMany({
+    const restaurant_menu = (await this.prisma.menu.findMany({
       where: {
         restaurantId: parseInt(restaurantId),
         deleted: false
@@ -219,7 +207,7 @@ export class MenuService {
         }
       ]
     }));
-    return new APIResponse(Status.Success, StatusMessage.GetSuccess, menus);
+    return restaurant_menu;
   }
 
   async getRestaurantMenuById(restaurantId: string, menuId: string) {
@@ -235,13 +223,10 @@ export class MenuService {
         }
       });
       if (!menu) {
-        return new APIResponse(Status.OtherErrors, StatusMessage.NoRecord, null);
+        throw new NotFoundException(StatusMessage.NoRecord);
       }
 
-      const base64Image = await this.ConvertToBase64String(menu.image)
-      menu.image = base64Image
-
-      return new APIResponse(Status.Success, StatusMessage.GetSuccess, menu);
+      return menu;
     } catch (error) {
       throw error
     }
@@ -259,12 +244,7 @@ export class MenuService {
           menuCategory: true,
         }
       });
-      for (let i = 0; i < menuList.length; i++) {
-        const base64Image = await this.ConvertToBase64String(menuList[i].image)
-        menuList[i].image = base64Image
-      }
-
-      return new APIResponse(Status.Success, StatusMessage.GetSuccess, menuList);
+      return  menuList;
     } catch (error) {
       throw error
     }
@@ -280,14 +260,14 @@ export class MenuService {
       }
     });
     if (!menu) {
-      return new APIResponse(Status.OtherErrors, StatusMessage.NoRecord, null);
+      throw new NotFoundException(StatusMessage.NoRecord);
     }
 
     if (menu.name.toLowerCase() == dto.name.toLowerCase()) {
-      return new APIResponse(Status.OtherErrors, StatusMessage.Exist, null);
+      throw new BadRequestException(StatusMessage.Exist)
     }
 
-    if (await fileExist(menu.image)) {
+    if (file && await fileExist(menu.image)) {
       await deleteFile(menu.image)
     }
 
@@ -313,7 +293,7 @@ export class MenuService {
       },
     });
 
-    this.socket.emitToClient(MenuManagementEvents.get_restaurant_menu_categories_event)
+    this.socket.emitToClient(MenuManagementEvents.get_restaurant_menu_event)
     return new APIResponse(Status.Success, StatusMessage.Updated, null);
   }
 
@@ -330,18 +310,19 @@ export class MenuService {
           deleted: true
         },
       });
+      this.socket.emitToClient(MenuManagementEvents.get_restaurant_menu_event)
       return new APIResponse(Status.Success, StatusMessage.Deleted, null);
     } catch (error) {
       throw error;
     }
   }
 
-  async ConvertToBase64String(path: string): Promise<string> {
+  // async ConvertToBase64String(path: string): Promise<string> {
 
-    const fileData = fs.readFileSync(path);
-    const base64String = base64js.fromByteArray(new Uint8Array(fileData));
+  //   const fileData = fs.readFileSync(path);
+  //   const base64String = base64js.fromByteArray(new Uint8Array(fileData));
 
-    return base64String;
-  }
+  //   return base64String;
+  // }
 
 }
